[Godfile data-store](2) extract sqlite-task-attempt-repository
Move the task and attempt persistence cluster out of
packages/data-store/src/sqlite-adapter.ts into
packages/data-store/src/sqlite-task-attempt-repository.ts over the
SqliteExecutor context introduced by slice 1. The adapter keeps every
moved method as a one-line delegate; behavior unchanged.
